// src/Components/type.ts
export type CommonCardConfigType = {
  image: string;
  description: string;
  title: string;
};

export interface CommunityData {
  ID: number;
  post_author: string;
  post_date: string;
  post_date_gmt: string;
  post_content: string;
  post_title: string;
  post_excerpt: string;
  post_status: string;
  comment_status: string;
  ping_status: string;
  post_password: string;
  post_name: string;
  to_ping: string;
  pinged: string;
  post_modified: string;
  post_modified_gmt: string;
  post_content_filtered: string;
  post_parent: number;
  guid: string;
  menu_order: number;
  post_type: string;
  post_mime_type: string;
  comment_count: string;
  filter: string;
  image_url: string;
  post_url: string;
}

export interface State {
  communities: CommunityData[];
  loading: boolean;
  error: string | null;
}